@@ -15,12 +15,8 @@ describe('Settings', () => {
 		cy.loginAdmin();
 		// Visit the home page
 		cy.visit('/');
-		// Open the sidebar if it is not already open
-		cy.get('[aria-label="Open sidebar"]').then(() => {
-			cy.get('button[id="sidebar-toggle-button"]').click();
-		});
-		// Click on the profile link
-		cy.get('button').contains(adminUser.name).click();
+		// Click on the user menu
+		cy.get('button[aria-label="User Menu"]').click();
 		// Click on the settings link
 		cy.get('button').contains('Settings').click();
 	});
@@ -29,19 +25,6 @@ describe('Settings', () => {
 		it('user can open the General modal and hit save', () => {
 			cy.get('button').contains('General').click();
 			cy.get('button').contains('Save').click();
-		});
-	});
-
-	context('Connections', () => {
-		it('user can open the Connections modal and hit save', () => {
-			cy.get('button').contains('Connections').click();
-			cy.get('button').contains('Save').click();
-		});
-	});
-
-	context('Models', () => {
-		it('user can open the Models modal', () => {
-			cy.get('button').contains('Models').click();
 		});
 	});
 
@@ -56,14 +39,6 @@ describe('Settings', () => {
 		it('user can open the Audio modal and hit save', () => {
 			cy.get('button').contains('Audio').click();
 			cy.get('button').contains('Save').click();
-		});
-	});
-
-	context('Images', () => {
-		it('user can open the Images modal and hit save', () => {
-			cy.get('button').contains('Images').click();
-			// Currently fails because the backend requires a valid URL
-			// cy.get('button').contains('Save').click();
 		});
 	});
 
