@@ -184,7 +184,7 @@
 							<div
 								class=" h-fit py-1 px-2 flex items-center justify-center rounded border border-black/10 capitalize text-gray-600 dark:border-white/10 dark:text-gray-300"
 							>
-								⌫
+								⌫/Delete
 							</div>
 						</div>
 					</div>
