@@ -1,6 +1,6 @@
 <script lang="ts">
 	export let className = 'w-4 h-4';
-	export let strokeWidth = '1.5';
+	export let strokeWidth = '2';
 </script>
 
 <svg
