@@ -10,7 +10,7 @@
 export const ssr = false;
 
 // How to manage the trailing slashes in the URLs
-// the URL for about page witll be /about with 'ignore' (default)
-// the URL for about page witll be /about/ with 'always'
+// the URL for about page will be /about with 'ignore' (default)
+// the URL for about page will be /about/ with 'always'
 // https://kit.svelte.dev/docs/page-options#trailingslash
 export const trailingSlash = 'ignore';
