@@ -1,12 +1,11 @@
 import adapter from '@sveltejs/adapter-static';
-import { vitePreprocess } from '@sveltejs/kit/vite';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
 	// for more information about preprocessors
 	preprocess: vitePreprocess(),
-
 	kit: {
 		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
@@ -17,8 +16,16 @@ const config = {
 			fallback: 'index.html'
 		})
 	},
+	vitePlugin: {
+		inspector: {
+			toggleKeyCombo: 'meta-shift', // Key combination to open the inspector
+			holdMode: false, // Enable or disable hold mode
+			showToggleButton: 'always', // Show toggle button ('always', 'active', 'never')
+			toggleButtonPos: 'bottom-right' // Position of the toggle button
+		}
+	},
 	onwarn: (warning, handler) => {
-		const { code, _ } = warning;
+		const { code } = warning;
 		if (code === 'css-unused-selector') return;
 
 		handler(warning);
